+++ tests/dbus/conn-requests.cpp
@@ -182,7 +182,7 @@ void TestConnRequests::initTestCase()
     QCOMPARE(mLoop->exec(), 0);
     QCOMPARE(mConn->isReady(), true);
 
-    QCOMPARE(mConn->status(), static_cast<uint>(Connection::StatusConnected));
+    QCOMPARE(mConn->status(), Connection::StatusConnected);
 
     QVERIFY(mConn->requestsInterface() != 0);
 }
