+++ tests/dbus/contacts-capabilities.cpp
@@ -114,7 +114,7 @@ void TestContactsCapabilities::initTestCase()
     QCOMPARE(mLoop->exec(), 0);
     QCOMPARE(mConn->isReady(), true);
 
-    QCOMPARE(mConn->status(), static_cast<uint>(Connection::StatusConnected));
+    QCOMPARE(mConn->status(), Connection::StatusConnected);
 
     QVERIFY(mConn->contactManager()->supportedFeatures().contains(Contact::FeatureCapabilities));
 }
